auth-service/requirements/base.txt
alembic==1.11.1
asyncpg==0.28.0
async-fastapi-jwt-auth==0.5.1
fastapi==0.95.2
fastapi-cache2[redis]==v0.2.1
fastapi-limiter==0.1.5
gunicorn==20.1.0
httptools==0.5.0
httpx==0.25.0
itsdangerous==2.1.2
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-instrumentation-fastapi==0.41b0
opentelemetry-exporter-jaeger==1.20.0
orjson==3.8.13
pydantic[email]==1.9.0
python-multipart==0.0.6
redis==4.4.2
sqlalchemy[asyncio]==1.4.49
typer[all]==0.9.0
uvicorn==0.23.2
uvloop==0.17.0 ; sys_platform != "win32" and implementation_name == "cpython"
werkzeug==2.3.6
